rename SourceFileLookup to SourceFileSelect
