--- src/QuickMarcEditor/SourceFileLookup/index.js
@@ -1,1 +1,0 @@
-export * from './SourceFileLookup';
